engine: extract esm/layers/iss_layer.js
